third_party/blink/web_tests/http/tests/websocket/close-on-unload-reference-in-parent-expected.txt

Test that WebSocket opened in iframe is closed when the iframe is unloaded even if the WebSocket is referred by the parent frame.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS ws on main document is ready.
PASS insert a iframe, where open ws called 'socket1'
PASS document.childWebSocket is non-null.
PASS 'socket1' is sent to the server. unload the iframe. expect receiving 'socket1' on ws...
PASS closedSocket is "socket1"
PASS closeCode is "1001"
PASS document.childWebSocket.readyState is 3
PASS successfullyParsed is true

TEST COMPLETE
